requirements.txt
asgiref==3.2.10
Django==3.1.1
django-crispy-forms==1.9.2
django-phonenumber-field==5.0.0
django-rest-registration==0.5.6
djangorestframework==3.12.2
phonenumbers==8.12.9

mccabe==0.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
sqlparse==0.3.1
typing==3.7.4.3
